Moving `parse_moddb_data` to launcher.downloader.moddb
